// src/api/generated/debArchive.schemas.ts
export interface S3Target {
  /** The AWS region for the S3 bucket. */
  region: string;
  /** The name of the S3 bucket. */
  bucket: string;
  /** Hostname used for S3 compatible storage.
  If set, the region is ignored. */
  endpoint?: string;
  /** AWS credentials (Access Key ID) to access the S3 bucket. */
  awsAccessKeyId: string;
  /** AWS credentials (Secret Access Key) to access the S3 bucket. */
  awsSecretAccessKey: string;
  /** Prefix used in the bucket.
  Defaults to the bucket root if not specified. */
  prefix?: string;
  /** ACL (Access Control List) used for published files. */
  acl?: string;
  /** The AWS S3 storage class to use. */
  storageClass?: string;
  /** Server-side encryption method. Defaults to none.
  Currently, the only available encryption method is `AES256`. */
  encryptionMethod?: string;
  /** Workaround for specific S3-compatible issues (e.g. "+" in filenames). */
  plusWorkaround?: boolean;
  /** Disables the `MultiDel` S3 API.
  Useful for S3-compatible cloud storage providers that do not support it. */
  disableMultiDel?: boolean;
  /** Disables Signature V4 support.
  Useful for S3-compatible cloud storage providers that do not support SigV4. */
  forceSigV2?: boolean;
  /** Enables debug logging for S3 operations. */
  debug?: boolean;
}

/**
 * SwiftTarget represents a Swift (OpenStack) storage destination.
 */
export interface SwiftTarget {
  /** The name of the container. */
  container: string;
  /** OpenStack credentials (Username) to access Keystone. */
  username: string;
  /** OpenStack credentials (Password) to access Keystone. */
  password: string;
  /** Prefix used in the container.
  Defaults to the container root if not specified. */
  prefix?: string;
  /** The full URL of the Keystone server (including port and version). */
  authUrl: string;
  /** The OpenStack tenant name. */
  tenant?: string;
  /** The OpenStack tenant ID. */
  tenantId?: string;
  /** The OpenStack domain name. */
  domain?: string;
  /** The OpenStack domain ID. */
  domainId?: string;
  /** The OpenStack domain name the tenant/project belongs to. */
  tenantDomain?: string;
  /** The OpenStack domain ID the tenant/project belongs to. */
  tenantDomainId?: string;
}

/**
 * PublicationTarget represents a storage destination (currently supports S3 or Swift).
 */
export interface PublicationTarget {
  name?: string;
  /** The user-defined display name. */
  displayName: string;
  /** The publication_id (uuid) is part of the `name`, and is exposed separately for convenience. */
  readonly publicationTargetId?: string;
  s3?: S3Target;
  swift?: SwiftTarget;
}